chore: Initiating the dance between order and chaos. Let the games begin
This commit establishes the foundational structure of the CheckpointTools
repository. It includes the addition of a MIT license file and a generic
.gitignore file that ensure that unnecessary files are not tracked by
version control. Additionally, a preliminary draft of the README.md has
been created to provide initial documentation about the project. Lastly,
the meson.build configuration is set up to compile a placeholder version
of the ckview executable, laying groundwork for future development.