chore: Update GitHub Actions runner to Ubuntu 24.04 and add compilation requirements
This commit updates the GitHub Actions workflow in 'release-meson.yml' by switching
from `ubuntu-22.04` to `ubuntu-24.04`. This change aims to ensure a more recent
environment with better support for modern C++ features, particularly addressing
the need for robust handling of the "<format>" header.

Additionally, an updated README.md now includes a 'Requirements' section that
explicitly states the necessity of using a compiler with mature support for
C++20.