feat: Display checkpoint metadata using colored tables in ckshow
This commit improves the visualization of checkpoint metadata in `ckshow`
by utilizing the same color and table system used for tensor information
display. As a result, running `ckshow -m <checkpointfile>` now presents
metadata that is aligned and colored.

Although this implementation is still preliminary, it demonstrates how
structured metadata can be presented in a more engaging format.